Partition and PSP install path support (#355)
Make the PSP install path (for eboot, ISO and PSX) configurable. This
enables support for plugins like gclite (just modify the configuration
to include the category folder). The newly added configuration is
treated as optional and will not be written unless set.

As a side-effect of altering this, make the vita partition configurable
so it has parity in the codebase.